Support handlers altering the set of files that appear in a directory listing.
In cases like .any.js files for wpt we create some virtual resources
for on-disk files. It would be nice if they appeared in directory
listings. To allow that to happen, we add a mechanism so that when a
handler is registered with the router we can also register a
corresponding listing_fixup function that takes the request and the
list of files in the directory and returns an amended list of files.